A NEW METHOD.
